refactor: modularizar layout principal
